pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-build-tool"
dynamic = ["version"]
description = "Infrastructure-grade prompt engineering for AI teams working across LLMs"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "PBT Team", email = "support@promptbuildtool.com"},
]
maintainers = [
    {name = "PBT Team", email = "support@promptbuildtool.com"},
]
keywords = [
    "prompt engineering",
    "llm",
    "ai",
    "machine learning",
    "claude",
    "openai",
    "gpt",
    "testing",
    "evaluation",
    "automation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "anthropic>=0.7.8",
    "openai>=1.3.7",
    "supabase>=2.0.2",
    "markdownify>=0.11.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "stripe>=7.8.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.22.0",
]
all = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "stripe>=7.8.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.1.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.urls]
Homepage = "https://promptbuildtool.com"
Documentation = "https://docs.promptbuildtool.com"
Repository = "https://github.com/prompt-build-tool/pbt"
"Bug Tracker" = "https://github.com/prompt-build-tool/pbt/issues"

[project.scripts]
pbt = "pbt.cli.main:app"
prompt-build-tool = "pbt.cli.main:app"

[tool.setuptools.dynamic]
version = {attr = "pbt.__version__.__version__"}

[tool.setuptools.packages.find]
include = ["pbt*"]

[tool.setuptools.package-data]
pbt = ["templates/*.yaml", "templates/*.sql", "static/*", "config/*.json"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]